Add a test from left over previous PR
Also: add another test that demonstrates some  behavior relating to default
parameters.

Also: Rename a test so that we don't accidentally get the test file.
In general, it's a bad idea to use a source file name in the compiler code
for a test since we often get the wrong file when loading it into the editor.