Replace sun.misc.Unsafe used in LazyVals with VarHandles (#24109)
This PR replaces the use of the deprecated Unsafe with VarHandles,
available in java 9+. Important thing about VarHandles is that for
private members (like our val dictating the value/resolution progress of
the lazy val), the findVarHandle method must be called from the class
containing the val, otherwise we end up with an error. This meant that
some custom logic had to be added to the MoveStatics phase, which would
usually move static constructor to the companion object (which we do not
want for the VarHandles here).
Only the newer implementation was adjusted, the one available under
`-Ylegacy-lazy-vals` was left untouched.