Use upper bound of abstract types in exhaustivity checking (#23909)
closes https://github.com/scala/scala3/issues/23620,
https://github.com/scala/scala3/issues/24246

If the upper bound of an abstract type is not `sealed` but is
effectively sealed, it is not handled correctly, since classSym could
return a type that is not `sealed` (Object in the issue above).
If the type were not abstract, it would pass the logic that checks
whether it is an Or, And, etc., and would be handled properly.
This PR makes exhaustivity checking use the upper bound of abstract
types that are effectively sealed.

---------

Co-authored-by: Zieliński Patryk <75637004+zielinsky@users.noreply.github.com>