Warn if type argument was inferred as union type (#24258)
Scala 2 has a `-Xlint:infer-any` option that warns if a type argument
was inferred as `Any`. In Scala 3 instead of union type arguments will
be inferred as union types. In Scala 3, a union type will be inferred
instead of Any. So to replace `infer-any` you need to warn that a union
type has been inferred.