Fix hole for recomputing denotations of NamedTypes
To recompute a denotation of a NamedType in a new phase we have a subtle logic
that sometimes recomputes the denotation from the prefix and the symbol or name,
and at other times recomputes it by transforming the previous denotation with
`current`. This logic was wrong for capturing types. We sometimes recomputed
by transforming the previous denotation where this led to wrong denotation.

We "fixed" problems this caused before by sometimes doing an explicit
`recomputeDenot`. But that's unsatisfactory since we can never know whether we do
that in all cases where it's needed.

This commit is a more principled fix where fix the logic that determines how
to recompute a NamedType's denotation by consulting sometimes the rechecker's advice.

 recomputed by transforming the previous denotation where this
led